Fixing build error for v1.9.x NCS
Changed:
* Fixing build warning related to atomic_get return value
* Fixing build warning related to fota_download_start

+++ lib/ble/ble_central.c
@@ -422,7 +422,7 @@ static void discovery_completed(struct bt_gatt_dm *dm, void *context)
         }
     }
 
-    LOG_DBG("Number of clients connected %i", atomic_get(&m_num_connected));
+    LOG_DBG("Number of clients connected %i", (int)atomic_get(&m_num_connected));
 
     // Set to ready
     atomic_inc(&m_num_connected);
